2nd complete version - User functionality
the transition to user stuff is complete, and i've uploaded all the seshs from April 13th till May 20th(today)

+++ hours_app/routers/crud_basic.py
@@ -5,49 +5,57 @@ from fastapi import FastAPI, Depends, Path, Query, Body, APIRouter
 from sqlmodel import Session, SQLModel, create_engine, Field, select, func
 from starlette.exceptions import HTTPException
 
-from ..dependencies import SessionDep
+from ..dependencies import SessionDep, oauth2_scheme
 from ..models import Sesh, SeshType, SeshCreate, SeshBase, SeshUpdate, UserInDB
 from .users import get_current_user
 
 
-router = APIRouter(tags=["crud"])
+router = APIRouter(tags=["crud"], dependencies=[Depends(oauth2_scheme)], prefix="/seshs")
 
 
-
-@router.get("/seshs/")
-def read_seshs(session: SessionDep):
-    seshs = session.exec(select(Sesh)).all()
+@router.get("/", summary="Read Seshs, clearly your own", dependencies=[Depends(oauth2_scheme)])
+def read_seshs(session: SessionDep, current_user: Annotated[UserInDB, Depends(get_current_user)]):
+    seshs = session.exec(select(Sesh).where(Sesh.owner_id == current_user.id)).all()
     return seshs
 
 
-@router.post("/seshs/")
+@router.post("/")
 def create_seshs(session: SessionDep, current_user: Annotated[UserInDB, Depends(get_current_user)],
-                 sesh_length: int, sesh_desc: str = "desc",  # i COULD add annotation with restrictions here # todoooooooooooooooo
+                 sesh_length: int, sub_tag_one: str = None, sub_tag_two: str = None,
+                 sesh_desc: str = None,  # i COULD add annotation with restrictions here # todoooooooooooooooo
                  sesh_type: SeshType = "programming",
-                 sesh_day: date = date.today()):
-    db_sesh = Sesh(length=sesh_length, specifics=sesh_desc, day=sesh_day, type=sesh_type, owner_id=current_user.id)
+                 sesh_day: date = date.today(),):
+    db_sesh = Sesh(length=sesh_length, specifics=sesh_desc, day=sesh_day, type=sesh_type, owner_id=current_user.id,
+                   sub_tag_one=sub_tag_one, sub_tag_two=sub_tag_two)
     session.add(db_sesh)
     session.commit()
     session.refresh(db_sesh)
     return db_sesh
 
 
-@router.delete("/seshs/")
-def delete_seshs(session: SessionDep, sesh_id: int):
+@router.delete("/")
+def delete_seshs(session: SessionDep, sesh_id: int, current_user: Annotated[UserInDB, Depends(get_current_user)]):
     sesh = session.get(Sesh, sesh_id)
     if not sesh:
         print("FUCKING KILL YOUSEFL")
         raise HTTPException(status_code=404, detail="FUCKING KILL YOURSELF")
+    if sesh.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not Your Sesh")
+
     session.delete(sesh)
     session.commit()
     return {"msg": "fuck off aye"}
 
 
-@router.put("/seshs/")
-def update_sesh(session: SessionDep, sesh: SeshUpdate, sesh_id: int):
+@router.put("/")
+def update_sesh(session: SessionDep, sesh: SeshUpdate,
+                sesh_id: int, current_user: Annotated[UserInDB, Depends(get_current_user)]):
     db_sesh = session.get(Sesh, sesh_id)
     if not db_sesh:
         raise HTTPException(status_code=404, detail="nope mate")
+    if db_sesh.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not Your Sesh")
+
     sesh_data = sesh.model_dump(exclude_unset=True)
     db_sesh.sqlmodel_update(sesh_data)
     session.add(db_sesh)
